requirements.txt
pyspellchecker>=0.7
textstat>=0.7.3
